refactor: Move and rework T4C repository user interface
