+++ backend/capellacollab/settings/modelsources/t4c/repositories/interface.py
@@ -57,8 +57,69 @@ def start_repository(instance: DatabaseT4CInstance, name: str):
 
 
 def stop_repository(instance: DatabaseT4CInstance, name: str):
-    r = get(
-        f"{instance.rest_api}/repositories/stop/{name}",
+    r = requests.get(
+        f"{instance.rest_api}/repositories/stop/{urllib.parse.quote(name, safe='')}",
+        auth=HTTPBasicAuth(instance.username, instance.password),
+        timeout=config["requests"]["timeout"],
+    )
+    r.raise_for_status()
+    return r.json()
+
+
+def add_user_to_repository(
+    instance: DatabaseT4CInstance,
+    repository_name: str,
+    username: str,
+    password: str = generate_password(),
+    is_admin: bool = False,
+):
+    r = requests.post(
+        f"{instance.rest_api}/users",
+        params={"repositoryName": repository_name},
+        json={
+            "id": username,
+            "isAdmin": is_admin,
+            "password": password,
+        },
+        auth=HTTPBasicAuth(instance.username, instance.password),
+        timeout=config["requests"]["timeout"],
+    )
+
+    # No exception if user does already exist (status_code 400)
+    if not r.ok and r.status_code != 400:
+        raise requests.HTTPError(r)
+    return r.json()
+
+
+def remove_user_from_repository(
+    instance: DatabaseT4CInstance, repository_name: str, username: str
+):
+    r = requests.delete(
+        f"{instance.rest_api}/users/{urllib.parse.quote(username, safe='')}",
+        config["modelsources"]["t4c"]["restAPI"] + "" + username,
+        params={"repositoryName": repository_name},
+        auth=HTTPBasicAuth(instance.username, instance.password),
+        timeout=config["requests"]["timeout"],
+    )
+    # No exception if user does not exist (status_code 404)
+    if not r.ok and r.status_code != 404:
+        raise requests.HTTPError(r)
+
+
+def update_password_of_user(
+    instance: DatabaseT4CInstance,
+    repository_name: str,
+    username: str,
+    password: str,
+):
+    r = requests.put(
+        f"{instance.rest_api}/users/{urllib.parse.quote(username, safe='')}",
+        params={"repositoryName": repository_name},
+        json={
+            "id": username,
+            "isAdmin": False,
+            "password": password,
+        },
         auth=HTTPBasicAuth(instance.username, instance.password),
         timeout=config["requests"]["timeout"],
     )
